Many changes....
Preliminary XRGB1555 implementation (with bugs of course :D)
ARGB8888 fix for games that didn't work, asteroid, starwars, others.... stella as well (but only half the screen is showing)
New audio backend, I did this for latency but I'm not certain of the results. I'll probably go back and try to figure out (again...) how to make it work there.
Support for core options!