Use Unity's Update(). Set material's texture only once.
Update() instead of InvokeRepeating() to demo time scaling, will probably move the update loop logic as part of the Wrapper.Update() function.

The need to check for texture update should not be needed (the wrapper is instantiated for every game so the format doesn't change), but applying the texture only at the start doesn't work on some cores for some reason.